\name{BFI228}
\alias{BFI228}
\docType{data}
\title{Ordinal Data of the Big Five Inventory (BFI)
}

\description{
The BFI228 is part of the study on personality and relationship satisfaction (Luo, 2005).  The participants were 228 undergraduate students at a large public university in the US.  The data were participants' self ratings on the 44 items of the Big Five Inventory (John, Donahue, & Kentle, 1991). These items are Likert variables: disagree strongly (1), disagree a little (2), neither agree nor disagree (3), agree a little (4), and agree strongly (5).

}

\usage{data(BFI228)}

\format{

The format is a n by p matrix of ordinal variables, where n is the number of participants (228) and p is the number of manifest variables (44).

}



\details{

The variables were ordered such that indicators of the same factor are grouped together. Note that  reverse-coded items are denoted by '_R'.

V01 to V08 are variables for the factor extraversion: talkative, reserved_R, fullenergy, enthusiastic, quiet_R, assertive, shy_R, and outgoing.

V09 to V17 are variables for the factor agreeableness: findfault_R, helpful, quarrels_R, forgiving, trusting, cold_R, considerate, rude_R, and cooperative.

V18 to V26 are variables for the factor conscientiousness are: thorough, careless_R, reliable, disorganized_R, lazy_R, persevere, efficient, plans, and distracted_R.

V27 to V34 are variables for the factor neuroticism: blue, relaxed_R, tense, worries, emostable_R, moody, calm_R, and nervous.

V35 to V44 are variables for the factor openness: ideas, curious, ingenious, imagination, inventive, artistic, routine_R, reflect, nonartistic, and sophisticated.

}


\references{
John, O. P., Donahue, E. M., & Kentle, R. L. (1991). The Big Five Inventory { versions 4a and 54. Berkeley, CA: University of California,Berkeley, Institute of Personality and Social Research.

Luo, S. (2005): unpublished study on personality traits and relationship satisfaction.

}
}

\keyword{datasets}
